requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
python-dotenv==1.0.0
elasticsearch==8.9.0
pytesseract==0.3.10
opencv-python==4.7.0.72
langchain==0.0.148
requests==2.28.1
beautifulsoup4==4.12.2
python-multipart
pdf2image
easyocr
langchain_community